refactor store-server and s3-proxy
