+++ s3-proxy/src/utils/type_utils.rs
